Data.Array

Description

Basic non-strict arrays.

Note: The Data.Array.IArray module provides a more general interface to immutable arrays: it defines operations with the same names as those defined below, but with more general types, and also defines Array instances of the relevant classes. To use that more general interface, import Data.Array.IArray but not Data.Array.

Immutable non-strict arrays

Haskell provides indexable arrays, which may be thought of as functions whose domains are isomorphic to contiguous subsets of the integers. Functions restricted in this way can be implemented efficiently; in particular, a programmer may reasonably expect rapid access to the components. To ensure the possibility of such an implementation, arrays are treated as data, not as general functions.

Since most array functions involve the class Ix, this module is exported from Data.Array so that modules need not import both Data.Array and Data.Ix.

module Data.Ix

data Array i e

Array construction

array :: Ix i => (i, i) -> [(i, e)] -> Array i e

listArray :: Ix i => (i, i) -> [e] -> Array i e

accumArray :: Ix i => (e -> a -> e) -> e -> (i, i) -> [(i, a)] -> Array i e

Accessing arrays

(!) :: Ix i => Array i e -> i -> e

bounds :: Array i e -> (i, i)

indices :: Ix i => Array i e -> [i]

elems :: Array i e -> [e]

assocs :: Ix i => Array i e -> [(i, e)]

Incremental array updates

(//) :: Ix i => Array i e -> [(i, e)] -> Array i e

accum :: Ix i => (e -> a -> e) -> Array i e -> [(i, a)] -> Array i e

Derived arrays

ixmap :: (Ix i, Ix j) => (i, i) -> (i -> j) -> Array j e -> Array i e