Data.Array.IO.Safe

Description

Mutable boxed and unboxed arrays in the IO monad. . Safe API only of Data.Array.IO.

Since: array-0.4.0.0

IO arrays with boxed elements

data IOArray i e

An IOArray is a mutable, boxed, non-strict array in the IO monad. The type arguments are as follows:

i: the index type of the array (should be an instance of Ix)

e: the element type of the array.

IO arrays with unboxed elements

data IOUArray i e

Mutable, unboxed, strict arrays in the IO monad. The type arguments are as follows:

i: the index type of the array (should be an instance of Ix)

e: the element type of the array. Only certain element types are supported: see Data.Array.MArray for a list of instances.

Overloaded mutable array interface

module Data.Array.MArray.Safe

Doing I/O with IOUArrays

hGetArray

Arguments

:: Handle
Handle to read from

-> IOUArray Int Word8
Array in which to place the values

-> Int
Number of Word8s to read

-> IO Int
Returns: the number of Word8s actually read, which might be smaller than the number requested if the end of file was reached.

Reads a number of Word8s from the specified Handle directly into an array.

hPutArray

Arguments

:: Handle
Handle to write to

-> IOUArray Int Word8
Array to write from

-> Int
Number of Word8s to write

-> IO ()

Writes an array of Word8 to the specified Handle.