Data.Array.MArray

Description

An overloaded interface to mutable arrays. For array types which can be used with this interface, see Data.Array.IO, Data.Array.ST, and Data.Array.Storable.

Class of mutable array types

class Monad m => MArray a e m where

Class of mutable array types.

An array type has the form (a i e) where a is the array type constructor (kind * -> * -> *), i is the index type (a member of the class Ix), and e is the element type.

The MArray class is parameterised over both a and e (so that instances specialised to certain element types can be defined, in the same way as for IArray), and also over the type of the monad, m, in which the mutable array will be manipulated.

Methods

getBounds :: Ix i => a i e -> m (i, i)

Returns the bounds of the array

newArray :: Ix i => (i, i) -> e -> m (a i e)

Builds a new array, with every element initialised to the supplied value.

newArray_ :: Ix i => (i, i) -> m (a i e)

Builds a new array, with every element initialised to an undefined value. In a monadic context in which operations must be deterministic (e.g. the ST monad), the array elements are initialised to a fixed but undefined value, such as zero.

The Ix class and operations

module Data.Ix

Constructing mutable arrays

newListArray :: (MArray a e m, Ix i) => (i, i) -> [e] -> m (a i e)

Constructs a mutable array from a list of initial elements. The list gives the elements of the array in ascending order beginning with the lowest index.

Reading and writing mutable arrays

readArray :: (MArray a e m, Ix i) => a i e -> i -> m e

Read an element from a mutable array

writeArray :: (MArray a e m, Ix i) => a i e -> i -> e -> m ()

Write an element in a mutable array

Derived arrays

mapArray :: (MArray a e' m, MArray a e m, Ix i) => (e' -> e) -> a i e' -> m (a i e)

Constructs a new array derived from the original array by applying a function to each of the elements.

mapIndices :: (MArray a e m, Ix i, Ix j) => (i, i) -> (i -> j) -> a j e -> m (a i e)

Constructs a new array derived from the original array by applying a function to each of the indices.

Deconstructing mutable arrays

getElems :: (MArray a e m, Ix i) => a i e -> m [e]

Return a list of all the elements of a mutable array

getAssocs :: (MArray a e m, Ix i) => a i e -> m [(i, e)]

Return a list of all the associations of a mutable array, in index order.

Conversions between mutable and immutable arrays

freeze :: (Ix i, MArray a e m, IArray b e) => a i e -> m (b i e)

Converts a mutable array (any instance of MArray) to an immutable array (any instance of IArray) by taking a complete copy of it.

unsafeFreeze :: (Ix i, MArray a e m, IArray b e) => a i e -> m (b i e)

Deprecated: Please import from Data.Array.Unsafe instead; This will be removed in the next release

Converts an mutable array into an immutable array. The implementation may either simply cast the array from one type to the other without copying the array, or it may take a full copy of the array.

Note that because the array is possibly not copied, any subsequent modifications made to the mutable version of the array may be shared with the immutable version. It is safe to use, therefore, if the mutable version is never modified after the freeze operation.

The non-copying implementation is supported between certain pairs of array types only; one constraint is that the array types must have identical representations. In GHC, The following pairs of array types have a non-copying O(1) implementation of unsafeFreeze. Because the optimised versions are enabled by specialisations, you will need to compile with optimisation (-O) to get them.

IOUArray -> UArray

STUArray -> UArray

IOArray -> Array

STArray -> Array

thaw :: (Ix i, IArray a e, MArray b e m) => a i e -> m (b i e)

Converts an immutable array (any instance of IArray) into a mutable array (any instance of MArray) by taking a complete copy of it.

unsafeThaw :: (Ix i, IArray a e, MArray b e m) => a i e -> m (b i e)

Deprecated: Please import from Data.Array.Unsafe instead; This will be removed in the next release

Converts an immutable array into a mutable array. The implementation may either simply cast the array from one type to the other without copying the array, or it may take a full copy of the array.

Note that because the array is possibly not copied, any subsequent modifications made to the mutable version of the array may be shared with the immutable version. It is only safe to use, therefore, if the immutable array is never referenced again in this thread, and there is no possibility that it can be also referenced in another thread. If you use an unsafeThawwriteunsafeFreeze sequence in a multi-threaded setting, then you must ensure that this sequence is atomic with respect to other threads, or a garbage collector crash may result (because the write may be writing to a frozen array).

The non-copying implementation is supported between certain pairs of array types only; one constraint is that the array types must have identical representations. In GHC, The following pairs of array types have a non-copying O(1) implementation of unsafeThaw. Because the optimised versions are enabled by specialisations, you will need to compile with optimisation (-O) to get them.

UArray -> IOUArray

UArray -> STUArray

Array -> IOArray

Array -> STArray