Data.Array.ST.Safe

Description

Mutable boxed and unboxed arrays in the ST monad.

Safe API only of Data.Array.ST.

Since: array-0.4.0.0

Boxed arrays

data STArray s i e

Mutable, boxed, non-strict arrays in the ST monad. The type arguments are as follows:

s: the state variable argument for the ST type

i: the index type of the array (should be an instance of Ix)

e: the element type of the array.

runSTArray :: (forall s. ST s (STArray s i e)) -> Array i e

A safe way to create and work with a mutable array before returning an immutable array for later perusal. This function avoids copying the array before returning it - it uses unsafeFreeze internally, but this wrapper is a safe interface to that function.

Unboxed arrays

data STUArray s i e

A mutable array with unboxed elements, that can be manipulated in the ST monad. The type arguments are as follows:

s: the state variable argument for the ST type

i: the index type of the array (should be an instance of Ix)

e: the element type of the array. Only certain element types are supported.

An STUArray will generally be more efficient (in terms of both time and space) than the equivalent boxed version (STArray) with the same element type. However, STUArray is strict in its elements - so don't use STUArray if you require the non-strictness that STArray provides.

runSTUArray :: (forall s. ST s (STUArray s i e)) -> UArray i e

A safe way to create and work with an unboxed mutable array before returning an immutable array for later perusal. This function avoids copying the array before returning it - it uses unsafeFreeze internally, but this wrapper is a safe interface to that function.

Overloaded mutable array interface

module Data.Array.MArray.Safe