Data.Array.Storable

Description

A storable array is an IO-mutable array which stores its contents in a contiguous memory block living in the C heap. Elements are stored according to the class Storable. You can obtain the pointer to the array contents to manipulate elements from languages like C.

It is similar to IOUArray but slower. Its advantage is that it's compatible with C.

Arrays of Storable things.

data StorableArray i e

The array type

Overloaded mutable array interface

Module Data.Array.MArray provides the interface of storable arrays. They are instances of class MArray (with the IO monad).

module Data.Array.MArray

Accessing the pointer to the array contents

withStorableArray :: StorableArray i e -> (Ptr e -> IO a) -> IO a

The pointer to the array contents is obtained by withStorableArray. The idea is similar to ForeignPtr (used internally here). The pointer should be used only during execution of the IO action retured by the function passed as argument to withStorableArray.

touchStorableArray :: StorableArray i e -> IO ()

If you want to use it afterwards, ensure that you touchStorableArray after the last use of the pointer, so the array is not freed too early.